msg_producer/requirements.txt
flask
oci
opentelemetry-api
opentelemetry-sdk